Temp fix for the splitter view issue where it didn't know what the width of the left pane should be on initially loading.
Once this thing schedules, looks nice, etc., it would be nice to rethink...everything. The splitter view could be part of a larger view,
and maybe that's how I should have done it to start with. For now, it's part of no view, but I can access it with CSS.